Real Estate
Constructed a class that will contain the data to be kept for a real
estate property. The property will be the base and will have some
pieces of information kept. I have demonstrated my command of the GUI
programming techniques. All program input is by means of a GUI, using
at least two different layout managers.